convert ints to size_t

+++ include/sgsmooth/SGSmooth.hpp
@@ -4,9 +4,9 @@
 #include <vector>
 
 // savitzky golay smoothing.
-std::vector<double> sg_smooth(const std::vector<double> &v, const int w, const int deg);
+std::vector<double> sg_smooth(const std::vector<double> &v, const size_t w, const int deg);
 //! numerical derivative based on savitzky golay smoothing.
-std::vector<double> sg_derivative(const std::vector<double> &v, const int w,
+std::vector<double> sg_derivative(const std::vector<double> &v, const size_t w,
                                 const int deg, const double h=1.0);
 
 #endif // __SGSMOOTH_HPP__
